+++ internal/org/organization.go
@@ -1,8 +1,8 @@
 package org
 
-import "github.com/c1emon/lemontree/model"
+import "github.com/c1emon/lemontree/pkg/gormx"
 
 type Organization struct {
-	model.BaseFields
+	gormx.BaseFields
 	Name string `json:"name" gorm:"column:name;type:varchar(256);uniqueIndex;not null"`
 }
