+++ internal/user/user.go
@@ -1,13 +1,13 @@
-package model
+package user
 
-import "github.com/c1emon/lemontree/model"
+import "github.com/c1emon/lemontree/pkg/gormx"
 
 type User struct {
-	model.BaseFields
+	gormx.BaseFields
 	Name           string `json:"name"`
 	OrganizationId string `json:"oid"`
 }
 
 type UserRepository interface {
-	model.BaseRepository[User]
+	gormx.BaseRepository[User]
 }
