+++ oidc/signin.go
@@ -1,9 +1,7 @@
 package oidc
 
-import (
-	"github.com/c1emon/lemontree/model"
-)
+import "github.com/c1emon/lemontree/pkg/gormx"
 
 type SigninRequest struct {
-	model.BaseFields
+	gormx.BaseFields
 }
