+++ pkg/gormx/base_model.go
@@ -1,4 +1,4 @@
-package model
+package gormx
 
 import (
 	"time"
