+++ pkg/gormx/base_repository.go
@@ -1,4 +1,4 @@
-package model
+package gormx
 
 import "context"
